Updated siglip_text/validate to work without the triton-embed-image deployment
Previously, these were together in a mono-repo so the validate text, which
came after the validate embed_image, utilized the running embed_image deployment.

Now with these split out, the code needed to be updated to load the image model in
validate.py.

Also added necessary things into the environment.yml that defines the development
environment for this repo.