Relaxed checks on truncation request parameter
Previously, only using True | False were allowed. This was done by checking truncation type and throwing error if not bool.  But this is more restrictive than what is accepted by the Huggingface Transformers.tokenizer.__call__() method.  So let's relax this requirement to allow some of the other value options to be acceptable.